refactor(web-serial): relocate SerialCommandService module (#546)
Made-with: Cursor
